+++ bitvm/src/hash/sha256.rs
@@ -913,11 +913,28 @@ pub fn maj(x: u32, y: u32, z: u32, stack_depth: u32) -> Script {
 
 #[cfg(test)]
 mod tests {
-    use crate::hash::blake3::push_bytes_hex;
     use crate::hash::sha256::*;
     use crate::treepp::{execute_script, script};
     use crate::u32::u32_std::{u32_equal, u32_equalverify};
     use sha2::{Digest, Sha256};
+   
+    fn push_bytes_hex(hex: &str) -> Script {
+        let hex: String = hex
+            .chars()
+            .filter(|c| c.is_ascii_digit() || c.is_ascii_alphabetic())
+            .collect();
+    
+        let bytes: Vec<u8> = (0..hex.len())
+            .step_by(2)
+            .map(|i| u8::from_str_radix(&hex[i..i + 2], 16).unwrap())
+            .collect::<Vec<u8>>();
+    
+        script! {
+            for byte in bytes.iter().rev() {
+                { *byte }
+            }
+        }
+    }
 
     fn rrot(x: u32, n: usize) -> u32 {
         if n == 0 {
